pyproject.toml
[project]
name = "class01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.9.0",
    "python-decouple>=3.8",
]
